VLCMediaSubcategoryViewController: add delegate
